requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.2
Deprecated==1.2.13
Django==4.0.1
django-ajax-datatable==4.4.4
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-timezone-field==5.0
django-treebeard==4.5.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
http-ece==1.1.0
idna==3.3
jellyfish==0.9.0
kombu==5.2.4
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.1
py-vapid==1.8.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyngrok==5.1.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pywebpush==1.14.0
PyYAML==6.0
redis==4.3.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
twilio==7.8.0
tzdata==2022.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
